- Added support for volume keys by switching layers through context key. - RAM optimizations - Fixed memory leak in ActionEvaluator - Added delay for startup of keyboard to avoid need of pressing reset button when setup method fails.

+++ firmware/VaKeyboard/lib/KeyboardSDK/src/Matrix/MatrixDebouncer.h
@@ -6,7 +6,7 @@ class MatrixDebouncer
 {
 private:
     uint8_t debounceTimeInMs = 2;
-    long** lastDebounceTimes;
+    uint8_t** lastDebounceTimes;
 
 public:
     MatrixDebouncer(IKeyboardDescriptor *keymapProvider, uint8_t debounceTimeInMs);
